requirements.txt
geographiclib             2.1
geopy                     2.4.1
holidays                  0.79
joblib                    1.5.1
jupyter_client            8.6.3
jupyter_core              5.8.1
keras                     3.11.3
matplotlib                3.10.3
numpy                     2.3.1
packaging                 25.0
pandas                    2.3.1
pillow                    11.2.1
pip                       25.2
plotly                    6.2.0
python-dateutil           2.9.0.post0
scikit-learn              1.7.1
scipy                     1.16.0
seaborn                   0.13.2
sklearn-preprocessing     0.1.0